requirements.txt
groq
python-telegram-bot
python-dotenv
tenacity
openpyxl
pytest
pytest-asyncio
aiomysql
cryptography
